Short deck on pointing stories as a survival technique
Short deck on why pointing stories is a survival technique. It's about saving energy and just (hopefully) surviving to the next day when you'll have more data. Congratulate yourself for surviving, examine the reasons why you survived, and keep moving. Story pointing is not intended to be an complete answer to anything. Most of the detail in this deck is in the delivery; the slides are just talking points.